activate deactivate impa

+++ src/pages/dashboard/info/impas.js
@@ -1,7 +1,7 @@
 
 'use client'
 import AdminLayout from "@/layouts/Admin/AdminLayout";
-import { lazy, useEffect, useState, useRef } from "react";
+import { useEffect, useState, useRef } from "react";
 import { DataTable } from 'primereact/datatable';
 import { Column } from 'primereact/column';
 import { Button } from "primereact/button";
@@ -14,6 +14,8 @@ import { setGridRowData } from "@/features/grid/gridSlice";
 import { InputText } from "primereact/inputtext";
 import { Toast } from "primereact/toast";
 import { useRouter } from "next/router";
+import { OverlayPanel } from 'primereact/overlaypanel';
+
 const dialogStyle = {
     marginTop: '10vh', // Adjust the top margin as needed
     display: 'flex',
@@ -25,8 +27,7 @@ const dialogStyle = {
 
 const Impas = () => {
     const [data, setData] = useState([]);
-    console.log('data')
-    console.log(data)
+    const op = useRef(null);
     const dispatch = useDispatch();
     const toast = useRef(null);
     const [searchTerm, setSearchTerm] = useState({
@@ -41,6 +42,7 @@ const Impas = () => {
     const [expandedRows, setExpandedRows] = useState(null);
     const [totalRecords, setTotalRecords] = useState(0);
     const [loading, setLoading] = useState(false);
+    const [selected, setSelected] = useState(false)
     const [lazyState, setlazyState] = useState({
         first: 0,
         rows: 10,
@@ -54,11 +56,13 @@ const Impas = () => {
 
 
     const allowExpansion = (rowData) => {
+        if (rowData.isActive === false) return;
         return rowData
     };
 
-    const handleFetch = async (action) => {
-        const {data} = await axios.post('/api/product/apiImpa', {
+    const handleFetch = async () => {
+        setLoading(true)
+        const { data } = await axios.post('/api/product/apiImpa', {
             action: 'findAll',
             skip: lazyState.first,
             limit: lazyState.rows,
@@ -66,22 +70,21 @@ const Impas = () => {
         })
         setData(data.result)
         setTotalRecords(data.totalRecords)
+        setLoading(false)
     }
+
+
+
     useEffect(() => {
-        if (searchTerm.greek) handleFetch('searchGreekImpa');
-        if (searchTerm.english) handleFetch('searchEng');
-        if (searchTerm.code) handleFetch('searchCode');
-        if (searchTerm.greek === '' && searchTerm.english === '' && searchTerm.code === '') {
-       
-            handleFetch('findAllWithProducts')
-        }
+        console.log('it should run')
+        handleFetch()
     }, [searchTerm, lazyState.rows, lazyState.first, submitted])
 
 
     const rowExpansionTemplate = (props) => {
         return (
-            <ExpandedDataTable 
-                id={props._id} 
+            <ExpandedDataTable
+                id={props._id}
                 setSubmitted={setSubmitted}
                 showSuccess={showSuccess}
                 showError={showError}
@@ -136,34 +139,80 @@ const Impas = () => {
         )
     }
 
+    const onDeactivate = async () => {
+        console.log('deactivate')
 
+        try {
+            let { data } = await axios.post('/api/product/apiImpa', { action: 'deactivate', selected: selected })
+            if (!data.success) return showError('Αποτυχία απενεργοποίησης')
+            setSelected([])
+            setSubmitted(prev => !prev)
+
+
+        } catch (e) {
+            showError('Προσπαθήστε ξανά')
+
+        }
+    }
+    const onActivate = async () => {
+        try {
+            let { data } = await axios.post('/api/product/apiImpa', { action: 'activate', selected: selected })
+            if (!data.success) return showError('Αποτυχία απενεργοποίησης')
+            setSelected([])
+            setSubmitted(prev => !prev)
+
+        } catch (e) {
+            showError('Προσπαθήστε ξανά')
+
+        }
+    }
     const LeftToolbarTemplate = () => {
         return (
             <div className="flex flex-wrap gap-2">
                 <Button label="Νέο" icon="pi pi-plus" severity="secondary" onClick={openNew} />
+                <div className="card flex justify-content-center">
+                    <Button icon="pi pi-angle-down" type="button" label="Impa Status" onClick={(e) => op.current.toggle(e)} />
+                    <OverlayPanel ref={op} >
+                        <div className="flex flex-column">
+                            <Button className="mb-2" icon="pi pi-times" label="Απενεργοποίηση Impa" severity="danger" onClick={onDeactivate} />
+                            <Button label="Eνεργοποίηση Impa" icon="pi pi-check" severity="success" onClick={onActivate} />
+                        </div>
+                    </OverlayPanel>
+                </div>
+
+
             </div>
         );
     };
 
     const ActionBodyTemplate = (rowData) => {
         return (
-            <Actions onEdit={() => onEdit(rowData)} onDelete={() => onDelete(rowData)} />
+            <Actions isActive={rowData.isActive} onEdit={() => onEdit(rowData)} onDelete={() => onDelete(rowData)} />
         )
     }
 
     const onEdit = (rowData) => {
+        if (rowData.isActive === false) return showError('Δεν μπορείτε να επεξεργαστείτε απενεργοποιημένο impa')
         setSubmitted(false);
         setEditDialog(true);
         dispatch(setGridRowData(rowData))
 
     }
-    const onDelete = async ({_id}) => {
-        let {data} = await axios.post('/api/product/apiImpa', {action: 'deleteOne', id: _id})
+    const onDelete = async ({ _id, isActive }) => {
+        if (isActive === false) return showError('Δεν μπορείτε διαγράψετε απενεργοποιημένο impa')
+        let { data } = await axios.post('/api/product/apiImpa', { action: 'deleteOne', id: _id })
 
         setSubmitted(true)
-        if(!data.success) showError('Αποτυχία Διαγραφής')
+        if (!data.success) showError('Αποτυχία Διαγραφής')
         showSuccess('Επιτυχής Διαγραφή')
     }
+
+    const onSelectionChange = (e) => {
+        setSelected(e.value)
+    }
+
+
+
 
     return (
         <AdminLayout>
@@ -171,29 +220,34 @@ const Impas = () => {
             <StepHeader text="Κωδικοί Impas" />
             <Toolbar start={LeftToolbarTemplate} ></Toolbar>
             <DataTable
-                showGridlines  
+                selectionMode={'checkbox'}
+                selection={selected}
+                onSelectionChange={onSelectionChange}
+                showGridlines
                 key={'code'}
                 value={data}
                 first={lazyState.first}
-                rows = {lazyState.rows}
+                rows={lazyState.rows}
                 onPage={onPage}
                 lazy
                 expandedRows={expandedRows}
                 onRowToggle={(e) => setExpandedRows(e.data)}
                 rowExpansionTemplate={rowExpansionTemplate}
-                paginator 
+                paginator
                 totalRecords={totalRecords}
                 loading={loading}
                 rowsPerPageOptions={[20, 50, 100, 200]}
                 filterDisplay="row"
                 tableStyle={{ minWidth: '50rem' }}
-                >
+            >
+                <Column selectionMode="multiple" filed="selection" headerStyle={{ width: '3rem' }}></Column>
                 <Column bodyStyle={{ textAlign: 'center' }} expander={allowExpansion} style={{ width: '20px' }} />
                 <Column field="code" header="code" filter filterElement={searchCode} showFilterMenu={false}></Column>
                 <Column field="englishDescription" header="Αγγλική Περιγραφή" filter filterElement={searchEngName} showFilterMenu={false}></Column>
-                <Column field="greekDescription" header="Ελληνική Περιγραφή"  filter filterElement={searchGreekName} showFilterMenu={false}></Column>
+                <Column field="greekDescription" header="Ελληνική Περιγραφή" filter filterElement={searchGreekName} showFilterMenu={false}></Column>
                 <Column field="unit" header="Unit"></Column>
-                <Column style={{width: '100px'}} body={ActionBodyTemplate}></Column>
+                <Column field="isActive" header="isActive" body={IsActive}></Column>
+                <Column style={{ width: '100px' }} body={ActionBodyTemplate}></Column>
             </DataTable>
             <EditDialog
                 style={dialogStyle}
@@ -221,26 +275,26 @@ const ExpandedDataTable = ({ id, setSubmitted, showSuccess, showError }) => {
     const [loading, setLoading] = useState(false);
     const [data, setData] = useState([]);
     const [refetch, setRefetch] = useState(false);
-  
+
     useEffect(() => {
         const handleFetch = async () => {
             setLoading(true);
-            let {data} = await axios.post('/api/product/apiImpa', {action: 'findImpaProducts', id: id})
+            let { data } = await axios.post('/api/product/apiImpa', { action: 'findImpaProducts', id: id })
             setData(data.result)
             setLoading(false);
-        
+
         }
         handleFetch()
     }, [refetch])
 
-    const renderHeader  =  () => {
+    const renderHeader = () => {
         const handleClick = () => {
             router.push(`/dashboard/products-to-impa/${id}`)
         }
         return (
             <div>
-            <Button label="Νέο" icon="pi pi-plus" severity="secondary"  onClick={handleClick}/>
-        </div>
+                <Button label="Νέο" icon="pi pi-plus" severity="secondary" onClick={handleClick} />
+            </div>
         )
     }
     const header = renderHeader()
@@ -250,8 +304,8 @@ const ExpandedDataTable = ({ id, setSubmitted, showSuccess, showError }) => {
 
         const handleDelete = async () => {
             setLoading(true);
-          let {data} = await axios.post('/api/product/apiImpa', {action: 'deleteImpaProduct', id: product._id, impaId: id})
-          if(!data.success) showError('Αποτυχία Διαγραφής')
+            let { data } = await axios.post('/api/product/apiImpa', { action: 'deleteImpaProduct', id: product._id, impaId: id })
+            if (!data.success) showError('Αποτυχία Διαγραφής')
             showSuccess('Επιτυχής Διαγραφή')
             setLoading(false);
             setRefetch(prev => !prev)
@@ -268,12 +322,12 @@ const ExpandedDataTable = ({ id, setSubmitted, showSuccess, showError }) => {
             <p className="font-semibold mb-3 ">Προϊόντα συσχετισμένα με impa:</p>
             <DataTable
                 loading={loading}
-                showGridlines   
-                 header={header}
-                dataKey="_id" 
+                showGridlines
+                header={header}
+                dataKey="_id"
                 value={data}>
                 <Column field="NAME" header="Προϊόν"></Column>
-                <Column field="CODE" style={{width: '50px'}} body={DeleteProducts}></Column>
+                <Column field="CODE" style={{ width: '50px' }} body={DeleteProducts}></Column>
 
             </DataTable>
         </div>
@@ -281,10 +335,17 @@ const ExpandedDataTable = ({ id, setSubmitted, showSuccess, showError }) => {
 }
 
 
+const IsActive = ({ isActive }) => {
+    return (
+        <div style={{ width: '20px', height: '20px' }} className={`${isActive ? "bg-green-500" : "bg-red-500"} border-round flex align-items-center justify-content-center`}>
+            {isActive ? <i className="pi pi-check text-white text-xs"></i> : <i className="pi pi-times text-white text-xs"></i>}
+        </div>
+    )
+}
 
 
+const Actions = ({ onEdit, onDelete, isActive }) => {
 
-const Actions = ({onEdit, onDelete}) => {
     return (
         <div>
             <i className="pi pi-pencil text-primary mr-3 cursor-pointer" onClick={onEdit}></i>
